Crate deploy

Modules

build_ctx

cfg

deb
Generates Debian packages (.deb)

docker

inst_ctx
Abstracts installing to directory, deb, etc

packages

pkg
Abstractions for package

registry

sel_ctx

systemd
This module implements support for systemd config generation Templates are in systemd/*, we just want to compile them and put to appropriate place

util

Structs

Params

Functions

build_jjs_components

build_testlib

compile_sample_contest

create_registry

env_add

generate_api_docs

generate_archive

generate_envscript

generate_json_schema

generate_man

generate_rustdoc

package